Fixed Padding when scrollbar Exists

+++ Prowl.Runtime/GUI/Layout/LayoutNode.API.cs
@@ -36,9 +36,9 @@
             return this;
         }
 
-        public LayoutNode Expand(int xOffset = 0, int yOffset = 0) => Width(Size.Percentage(1f, xOffset)).Height(Size.Percentage(1f, yOffset));
-        public LayoutNode ExpandWidth(int pixelOffset = 0) => Width(Size.Percentage(1f, pixelOffset));
-        public LayoutNode ExpandHeight(int pixelOffset = 0) => Height(Size.Percentage(1f, pixelOffset));
+        public LayoutNode Expand(double xOffset = 0, double yOffset = 0) => Width(Size.Percentage(1f, xOffset)).Height(Size.Percentage(1f, yOffset));
+        public LayoutNode ExpandWidth(double pixelOffset = 0) => Width(Size.Percentage(1f, pixelOffset));
+        public LayoutNode ExpandHeight(double pixelOffset = 0) => Height(Size.Percentage(1f, pixelOffset));
 
         public LayoutNode Scale(Size size) => Width(size).Height(size);
         public LayoutNode Scale(Size width, Size height) => Width(width).Height(height);
